python/requirement.txt
mysql-connector-python==8.0.21
Flask==1.1.2
faker==4.1.1